backend/pyproject.toml
[project]
name = "sentinel-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "chromadb>=0.5.0",
    "llama-index>=0.10.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "openai>=1.30.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "uagents>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
